+++ Backend/src/main/java/com/AlaCartApp/controller/ProductController.java
@@ -2,7 +2,6 @@ package com.AlaCartApp.controller;
 
 import com.AlaCartApp.models.response.ProductDto;
 import com.AlaCartApp.service.abstraction.ProductService;
-import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,6 +14,7 @@ import java.util.List;
 @RequestMapping("/products")
 @RestController
 public class ProductController {
+
 
     private final ProductService productService;
 
@@ -33,8 +33,17 @@ public class ProductController {
                 .build();
     }
 
+    @GetMapping("/find-category-list")
+    public ResponseEntity<?> findByCategory(@RequestBody Long categoryId){
+        if (categoryId.equals(null) || categoryId < 0) return new ResponseEntity<>("FIELD NOT SUPPORTED", HttpStatus.BAD_REQUEST);
+        List<ProductDto> products = productService.findByCategory(categoryId);
+        if(products.isEmpty()) return new ResponseEntity<>("Not found products", HttpStatus.NO_CONTENT);
+        else return new ResponseEntity<>(products,HttpStatus.OK);
+    }
+
     @GetMapping("/{id}")
-    public ResponseEntity<ProductDto> findById(@PathVariable Long id){
+    public ResponseEntity<?> findById(@PathVariable Long id){
+        if (id<0|| id.equals(null)) return new ResponseEntity<>("FIELD NOT SUPPORTED", HttpStatus.BAD_REQUEST);
         return ResponseEntity.of(productService.findById(id));
     }
 
@@ -45,7 +54,8 @@ public class ProductController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Boolean> delete(@PathVariable Long id){
+    public ResponseEntity<?> delete(@PathVariable Long id){
+        if (id<0|| id.equals(null)) return new ResponseEntity<>("FIELD NOT SUPPORTED", HttpStatus.BAD_REQUEST);
         return new ResponseEntity<>(productService.delete(id)?HttpStatus.OK:HttpStatus.NOT_FOUND);
     }
 }
